+++ Number_Node.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 #include "Number_Node.h"
 
-Number_Node::Number_Node(int n)
+Number_Node::Number_Node(double n)
 :
 val(n)
 {
@@ -17,7 +17,7 @@ void Number_Node::accept(Expr_Node_Visitor& v)
     v.Visit_Number_Node(*this);
 }
 
-int Number_Node::getVal()
+double Number_Node::getVal()
 {
     return val;
 }
