+++ Number_Node.h
@@ -8,13 +8,13 @@
 class Number_Node : public Expr_Node
 {
     public:
-        Number_Node(int n);
+        Number_Node(double n);
         virtual ~Number_Node();
         virtual void accept (Expr_Node_Visitor &v);
-        int getVal();
+        double getVal();
     protected:
     private:
-        int val;
+        double val;
 };
 
 #endif // NUMBER_NODE_H
